google_scholar_crawler/requirements.txt
jsonpickle==1.4.2
scholarly==1.5.1
PyGithub==1.56